Add initial bot functionality
This is probably the biggest step we've recently made on this project's
side of things. The bot is mostly functional, but not thoroughly
bug-tested at all. From here, there are several issues we should
address:

- Add exception handling and safety. We currently don't check nullable
fields for our json objects; this has to change.

- Tidy up our internal flags for managing state; we should explicitly
decide on where to set our starting flags, and where the flags should
change in a consistent manner

- Add documentation for allowing overriding callback functionality via
inheritance. We should also ask: is this desired behavior? Maybe we
should find some alternative that's still compatible with private class
members, or evaluate whether we really need that safety (hint: we
probably do).

- Flesh out event handling and adding more features; we want to be
full-featured eventually, right? :)